refactor caches of isMsl() function
