+++ gradle.properties
@@ -15,4 +15,4 @@ propertiesPluginEnvironmentNameProperty=shortPlatformVersion
 # properties files
 # supported versions are 221, 222, 223, default is 221
 # pass ORG_GRADLE_PROJECT_shortPlatformVersion environment variable to overwrite
-shortPlatformVersion=221
+shortPlatformVersion=223
